feat(mlop-2760): moving writer_metadata to a new module
